# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Product Foundation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/session/organization": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Switch workspace",
        "description": "Only your own workspaces; others look like they don't exist (404).",
        "operationId": "switch_organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/organizations": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "My workspaces",
        "description": "Only workspaces you are a member of.",
        "operationId": "list_organizations",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationRead"
                  },
                  "type": "array",
                  "title": "Response List Organizations"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Create a workspace",
        "description": "You become its owner, and it becomes your active workspace.",
        "operationId": "create_organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationRead"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/current": {
      "patch": {
        "tags": [
          "organizations"
        ],
        "summary": "Rename the active workspace",
        "description": "Owners and admins.",
        "operationId": "update_organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationRead"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/current/leave": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Leave the active workspace",
        "description": "You lose access. The owner can't leave (transfer ownership first).",
        "operationId": "leave_organization",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/current/transfer-ownership": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Make another member the owner",
        "description": "Owner only. You become an admin.",
        "operationId": "transfer_ownership",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnership"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberList"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/current/deletion": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Schedule deleting the active workspace",
        "description": "Owner only. Type the workspace name to confirm. Deleted after the grace period.",
        "operationId": "schedule_organization_deletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationRead"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "organizations"
        ],
        "summary": "Cancel deleting the active workspace",
        "description": "Owner only.",
        "operationId": "cancel_organization_deletion",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/current/exports": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Export all data of the active workspace (ZIP)",
        "description": "Owners and admins. Follow the job, then download the ZIP.",
        "operationId": "export_organization",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRead"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency is not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeletionRequest": {
        "properties": {
          "confirm": {
            "type": "string",
            "maxLength": 320,
            "minLength": 1,
            "title": "Confirm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirm"
        ],
        "title": "DeletionRequest",
        "description": "Type the confirmation text exactly (the email address, or the workspace name)."
      },
      "DeletionStatus": {
        "properties": {
          "deletion_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletion Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "deletion_scheduled_at"
        ],
        "title": "DeletionStatus",
        "description": "When the deletion will happen (null = not scheduled)."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "description": "What went wrong."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Error envelope: `{\"error\": {...}}`."
      },
...
      "ExportRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ExportScope"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id",
            "description": "Follow progress with GET /api/jobs/{id}."
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "job_id",
          "filename",
          "ready",
          "size_bytes",
          "created_at",
          "expires_at",
          "download_url"
        ],
        "title": "ExportRead",
        "description": "A data export. Download it when `ready` is true."
      },
      "ExportScope": {
        "type": "string",
        "enum": [
          "account",
          "organization"
        ],
        "title": "ExportScope",
        "description": "Whose data is in the export."
      },
...
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrganizationRead"
            },
            "type": "array",
            "title": "Organizations"
          },
          "active_organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Active Organization Id"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "type": "array",
            "title": "Permissions",
            "description": "What you may do in the active workspace (the UI hides everything else)."
          }
        },
        "type": "object",
        "required": [
          "user",
          "organizations",
          "active_organization_id",
          "permissions"
        ],
        "title": "MeResponse",
        "description": "Everything the app needs after sign-in."
      },
      "MemberList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemberRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MemberList",
        "description": "Owner first, then admins, then members."
      },
      "MemberRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          },
          "is_you": {
            "type": "boolean",
            "title": "Is You"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "email",
          "role",
          "joined_at",
          "is_you"
        ],
        "title": "MemberRead",
        "description": "A person in the workspace."
      },
...
      "OrganizationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationCreate",
        "description": "Create a workspace."
      },
      "OrganizationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "deletion_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletion Scheduled At",
            "description": "Set when the workspace will be deleted (can be cancelled)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "OrganizationRead",
        "description": "An organization and the current user's role in it."
      },
      "OrganizationUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationUpdate",
        "description": "Rename the active workspace."
      },
      "Permission": {
        "type": "string",
        "enum": [
          "org:update",
          "org:delete",
          "org:export",
          "ownership:transfer",
          "members:read",
          "members:invite",
          "members:manage",
          "api_keys:manage",
          "audit:read",
          "jobs:read",
          "jobs:write"
        ],
        "title": "Permission",
        "description": "Something a caller may be allowed to do in a workspace."
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member"
        ],
        "title": "Role",
        "description": "What a member may do. The rules live in app/core/permissions.py."
      },
...
      "SwitchOrganization": {
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "organization_id"
        ],
        "title": "SwitchOrganization",
        "description": "Switch the active workspace for this browser."
      },
...
      "TransferOwnership": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "TransferOwnership",
        "description": "Make another member the owner. You become an admin."
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          },
          "has_password": {
            "type": "boolean",
            "title": "Has Password"
          },
          "google_linked": {
            "type": "boolean",
            "title": "Google Linked"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "deletion_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletion Scheduled At",
            "description": "Set when the account will be deleted (can be cancelled)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "email_verified",
          "has_password",
          "google_linked",
          "created_at"
        ],
        "title": "UserRead",
        "description": "The signed-in user."
      }
    }
  }
}